+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All important changes to `fiisoft-money` will be documented in this file
 
+## 1.1.0
+
+Two new public methods in MoneyFunc: normaliseStringAmount() and formatAmountAsString().
+File composer.lock removed from project.
+
 ## 1.0.0
 
 First version of library.